Move class_phpbb_functions into class_qi
