+++ modules/qi_create.php
@@ -34,7 +34,7 @@ class qi_create
 		// phpBB 3.2.0-3.2.1 is not compat with PHP 7.2 (702000)
 		// phpBB 3.2.x is not compat with PHP 7.3 (703000)
 		if ((PHP_VERSION_ID >= 70000 && !defined('PHPBB_32')) ||
-			(PHP_VERSION_ID >= 70200 && phpbb_version_compare(PHPBB_VERSION, '3.2.2', '<')) ||
+			(PHP_VERSION_ID >= 70200 && qi::phpbb_version_compare(PHPBB_VERSION, '3.2.2', '<')) ||
 			(PHP_VERSION_ID >= 70300 && !defined('PHPBB_33'))
 		)
 		{
